move views into correct directories
