+++ web-client/src/views/AppComponent.jsx
@@ -6,8 +6,8 @@ import { BeforeStartingCase } from './BeforeStartingCase';
 import { BeforeYouFileADocument } from './FileDocument/BeforeYouFileADocument';
 import { BlockedCasesReport } from './BlockedCasesReport/BlockedCasesReport';
 import { CaseDeadlines } from './CaseDeadlines/CaseDeadlines';
-import { CaseDetail } from './CaseDetail';
-import { CaseDetailInternal } from './CaseDetailInternal';
+import { CaseDetail } from './CaseDetail/CaseDetail';
+import { CaseDetailInternal } from './CaseDetail/CaseDetailInternal';
 import { CaseSearchNoMatches } from './CaseSearchNoMatches';
 import { CourtIssuedDocketEntry } from './CourtIssuedDocketEntry/CourtIssuedDocketEntry';
 import { CreateOrder } from './CreateOrder/CreateOrder';
