+++ web-client/src/views/CaseDetail/CaseDetailInternal.jsx
@@ -1,26 +1,26 @@
-import { AddToTrialModal } from './CaseDetail/AddToTrialModal';
-import { BlockFromTrialModal } from './CaseDetail/BlockFromTrialModal';
-import { CaseDeadlinesInternal } from './CaseDetail/CaseDeadlinesInternal';
+import { AddToTrialModal } from './AddToTrialModal';
+import { BlockFromTrialModal } from './BlockFromTrialModal';
+import { CaseDeadlinesInternal } from './CaseDeadlinesInternal';
 import { CaseDetailHeader } from './CaseDetailHeader';
-import { CaseDetailPendingReportList } from './CaseDetail/CaseDetailPendingReportList';
+import { CaseDetailPendingReportList } from './CaseDetailPendingReportList';
 import { CaseDetailSubnavTabs } from './CaseDetailSubnavTabs';
-import { CaseInformationInternal } from './CaseDetail/CaseInformationInternal';
-import { CaseNotes } from './CaseDetail/CaseNotes';
-import { CreateCaseDeadlineModalDialog } from './CaseDetail/CreateCaseDeadlineModalDialog';
-import { DeleteCaseDeadlineModalDialog } from './CaseDetail/DeleteCaseDeadlineModalDialog';
-import { DocketRecord } from './DocketRecord/DocketRecord';
-import { DraftDocuments } from './DraftDocuments/DraftDocuments';
-import { EditCaseDeadlineModalDialog } from './CaseDetail/EditCaseDeadlineModalDialog';
-import { ErrorNotification } from './ErrorNotification';
+import { CaseInformationInternal } from './CaseInformationInternal';
+import { CaseNotes } from './CaseNotes';
+import { CreateCaseDeadlineModalDialog } from './CreateCaseDeadlineModalDialog';
+import { DeleteCaseDeadlineModalDialog } from './DeleteCaseDeadlineModalDialog';
+import { DocketRecord } from '../DocketRecord/DocketRecord';
+import { DraftDocuments } from '../DraftDocuments/DraftDocuments';
+import { EditCaseDeadlineModalDialog } from './EditCaseDeadlineModalDialog';
+import { ErrorNotification } from '../ErrorNotification';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { MessagesInProgress } from './CaseDetail/MessagesInProgress';
-import { PartyInformation } from './CaseDetail/PartyInformation';
-import { PrioritizeCaseModal } from './CaseDetail/PrioritizeCaseModal';
-import { RemoveFromTrialSessionModal } from './CaseDetail/RemoveFromTrialSessionModal';
-import { SuccessNotification } from './SuccessNotification';
-import { Tab, Tabs } from '../ustc-ui/Tabs/Tabs';
-import { UnblockFromTrialModal } from './CaseDetail/UnblockFromTrialModal';
-import { UnprioritizeCaseModal } from './CaseDetail/UnprioritizeCaseModal';
+import { MessagesInProgress } from './MessagesInProgress';
+import { PartyInformation } from './PartyInformation';
+import { PrioritizeCaseModal } from './PrioritizeCaseModal';
+import { RemoveFromTrialSessionModal } from './RemoveFromTrialSessionModal';
+import { SuccessNotification } from '../SuccessNotification';
+import { Tab, Tabs } from '../../ustc-ui/Tabs/Tabs';
+import { UnblockFromTrialModal } from './UnblockFromTrialModal';
+import { UnprioritizeCaseModal } from './UnprioritizeCaseModal';
 import { connect } from '@cerebral/react';
 import { state } from 'cerebral';
 import React from 'react';
