+++ web-client/src/views/RequestAccess/RequestAccessWizard.jsx
@@ -1,4 +1,4 @@
-import { CaseDetailHeader } from '../CaseDetailHeader';
+import { CaseDetailHeader } from '../CaseDetail/CaseDetailHeader';
 import { ErrorNotification } from '../ErrorNotification';
 import { FormCancelModalDialog } from '../FormCancelModalDialog';
 import { RequestAccess } from './RequestAccess';
