+++ web-client/src/views/SignStipDecision.jsx
@@ -1,5 +1,5 @@
 import { Button } from '../ustc-ui/Button/Button';
-import { CaseDetailHeader } from './CaseDetailHeader';
+import { CaseDetailHeader } from './CaseDetail/CaseDetailHeader';
 import { PDFSignerMessage } from './PDFSignerMessage';
 import { PDFSignerToolbar } from './PDFSignerToolbar';
 import { connect } from '@cerebral/react';
